requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
python-jose
pydantic[email]
python-multipart
